Create AbstractSyntaxTreeResolver
