+++ core/src/main/scala/io/github/scalaquest/core/pipeline/resolver/AbstractSyntaxTreeResolver.scala
@@ -1,0 +1,39 @@
+package io.github.scalaquest.core.pipeline.resolver
+
+import io.github.scalaquest.core.dictionary.verbs.VerbPrep
+import io.github.scalaquest.core.model.{Action, ItemDescription, ItemRef}
+import io.github.scalaquest.core.pipeline.parser.{AbstractSyntaxTree, ParserResult}
+
+abstract class AbstractSyntaxTreeResolver extends Resolver {
+
+  def actions(v: VerbPrep): Either[String, Action]
+
+  def items(d: ItemDescription): Either[String, ItemRef]
+
+  override def resolve(parserResult: ParserResult): Either[String, ResolverResult] = {
+    for {
+      statement <- parserResult.tree match {
+        case AbstractSyntaxTree.Intransitive(verb, prep, _) =>
+          for {
+            action <- actions((verb, prep))
+          } yield Statement.Intransitive(action)
+
+        case AbstractSyntaxTree.Transitive(verb, prep, _, obj) =>
+          for {
+            action  <- actions((verb, prep))
+            itemRef <- items(obj)
+          } yield Statement.Transitive(action, itemRef)
+
+        case AbstractSyntaxTree.Ditransitive(verb, prep, _, directObj, indirectObj) =>
+          for {
+            action          <- actions((verb, prep))
+            directItemRef   <- items(directObj)
+            indirectItemRef <- items(indirectObj)
+          } yield Statement.Ditransitive(action, directItemRef, indirectItemRef)
+
+        case _ => Left("The statement is wrong.")
+      }
+    } yield ResolverResult(statement)
+  }
+
+}
